+++ spec/mongoid/paranoia_spec.rb
@@ -2,10 +2,78 @@
 require 'spec_helper'
 
 describe 'Mongoid::Paranoia with Mongoid::Slug' do
+  let(:paranoid_doc)    { ParanoidDocument.create!(title: 'slug') }
+  let(:paranoid_doc_2)  { ParanoidDocument.create!(title: 'slug') }
+  let(:paranoid_perm)   { ParanoidPermanent.create!(title: 'slug') }
+  let(:paranoid_perm_2) { ParanoidPermanent.create!(title: 'slug') }
+  let(:non_paranoid_doc) { Article.create!(title: 'slug') }
+  subject { paranoid_doc }
+
+  describe '.paranoid?' do
+    context 'when Mongoid::Paranoia is included' do
+      subject { paranoid_doc.class }
+      specify { expect(subject.is_paranoid_doc?).to be true }
+    end
+
+    context 'when Mongoid::Paranoia not included' do
+      subject { non_paranoid_doc.class }
+      specify { expect(subject.is_paranoid_doc?).to be false }
+    end
+  end
+
+  describe '#paranoid_deleted?' do
+    context 'when Mongoid::Paranoia is included' do
+      context 'when not destroyed' do
+        specify { expect(subject.paranoid_deleted?).to be false }
+      end
+
+      context 'when destroyed' do
+        before { subject.destroy }
+        specify { expect(subject.paranoid_deleted?).to be true }
+      end
+    end
+
+    context 'when Mongoid::Paranoia not included' do
+      subject { non_paranoid_doc }
+      specify { expect(subject.paranoid_deleted?).to be false }
+    end
+  end
+
+  describe 'restore callbacks' do
+    context 'when Mongoid::Paranoia is included' do
+      subject { paranoid_doc.class }
+      it { is_expected.to respond_to(:before_restore) }
+      it { is_expected.to respond_to(:after_restore) }
+    end
+
+    context 'when Mongoid::Paranoia not included' do
+      it { is_expected.to_not respond_to(:before_restore) }
+      it { is_expected.to_not respond_to(:after_restore) }
+    end
+  end
+
+  describe 'index' do
+    context 'simple index' do
+      before  { ParanoidDocument.create_indexes }
+      after   { ParanoidDocument.remove_indexes }
+      subject { ParanoidDocument }
+
+      it_should_behave_like 'has an index', { _slugs: 1 }, unique: true, sparse: true
+    end
+
+    context 'compound index' do
+      before  { ParanoidPermanent.create_indexes }
+      after   { ParanoidPermanent.remove_indexes }
+      subject { ParanoidPermanent }
+
+      it_should_behave_like 'has an index', { foo: 1, _slugs: 1 }, unique: nil, sparse: nil
+    end
+  end
+
   shared_examples_for 'paranoid slugs' do
     context 'querying' do
       it 'returns paranoid_doc for correct slug' do
-        expect(subject.class.find(subject.slug)).to eq(subject)
+        expect(subject.class.find(subject.slug)).to eq subject
       end
     end
 
@@ -95,30 +163,30 @@ describe 'Mongoid::Paranoia with Mongoid::Slug' do
       describe '.paranoid?' do
         context 'when Mongoid::Paranoia is included' do
           subject { paranoid_doc.class }
-          its(:is_paranoid_doc?) { should be_truthy }
+          its(:is_paranoid_doc?) { is_expected.to be_truthy }
         end
 
         context 'when Mongoid::Paranoia not included' do
           subject { non_paranoid_doc.class }
-          its(:is_paranoid_doc?) { should be_falsey }
+          its(:is_paranoid_doc?) { is_expected.to be_falsey }
         end
       end
 
       describe '#paranoid_deleted?' do
         context 'when Mongoid::Paranoia is included' do
           context 'when not destroyed' do
-            its(:paranoid_deleted?) { should be_falsey }
+            its(:paranoid_deleted?) { is_expected.to be_falsey }
           end
 
           context 'when destroyed' do
             before { subject.destroy }
-            its(:paranoid_deleted?) { should be_truthy }
+            its(:paranoid_deleted?) { is_expected.to be_truthy }
           end
         end
 
         context 'when Mongoid::Paranoia not included' do
           subject { non_paranoid_doc }
-          its(:paranoid_deleted?) { should be_falsey }
+          its(:paranoid_deleted?) { is_expected.to be_falsey }
         end
       end